anchored overlay can take an external ref for the anchor
