+++ src/hooks/useOnOutsideClick.tsx
@@ -14,24 +14,6 @@ type ShouldCallClickHandlerSettings = {
   e: TouchOrMouseEvent
 }
 
-const handlers: ((e: MouseEvent) => void)[] = []
-
-/**
- * Calls all handlers in reverse order
- * @param event The MouseEvent generated by the click event.
- */
-function handleClick(event: MouseEvent) {
-  if (!event.defaultPrevented) {
-    for (let i = handlers.length - 1; i >= 0; --i) {
-      handlers[i](event)
-      // eslint-disable-next-line @typescript-eslint/no-unnecessary-condition
-      if (event.defaultPrevented) {
-        break
-      }
-    }
-  }
-}
-
 const shouldCallClickHandler = ({ignoreClickRefs, containerRef, e}: ShouldCallClickHandlerSettings): boolean => {
   let shouldCallHandler = true
 
@@ -65,19 +47,12 @@ export const useOnOutsideClick = ({containerRef, ignoreClickRefs, onClickOutside
     },
     [onClickOutside, containerRef, ignoreClickRefs]
   )
+
   useEffect(() => {
-    if (handlers.length === 0) {
-      document.addEventListener('mousedown', handleClick)
-    }
-    handlers.push(onOutsideClickInternal)
+    // use capture to ensure we get all events
+    document.addEventListener('mousedown', onOutsideClickInternal, {capture: true})
     return () => {
-      handlers.splice(
-        handlers.findIndex(h => h === onOutsideClickInternal),
-        1
-      )
-      if (handlers.length === 0) {
-        document.removeEventListener('mousedown', handleClick)
-      }
+      document.removeEventListener('mousedown', onOutsideClickInternal, {capture: true})
     }
   }, [onOutsideClickInternal])
 }
